Open Telemetry Integrations (#7)
* adjusted telemetry

adjusted the telemetry data capture to associate single activities for the entire process and add events inside for specific occurences as well as extended the Context for Middleware to include the Activity if it exists.

* finalized OTEL design

finalized the OTEL design and linkage as well as implemented a test in one sample and updated the sample to include a BulkPublish call.

* adding unit testing

adding initial unit testing for Otel data

* expanded tests

Expanded unit testing to include otel mapped connection testing and added in a tag for the connection name

* completed unit testing

completed unit testing for the Open Telemetry

* updating references

updated the referenced libraries for the given connectors

* Update Shared.props

updating version number

* code cleanup

updating code post cleanup