3.0 (#18)
* migrated attributes

migrated the message and consumer attributes into a simpler set where it's now a single attribute for each instead of each aspect to produce a cleaner codebase.

* shifted consumer connection returns

shifted the consumer connection returns in order to allow for an extension to create fluent type calls.

* added extension for consumers

added in an extension to allow for consumers to be registered in a fluent style.

* tweaks for Kafka

tweaks in the Kafka system to handle empty reply topics when partition setting fails to align with the reply topic properly

* code cleanup

cleaning up some code that was used for testing purposes

* updating generic names

updated generic names for ease of reading/code quality

* adding initial CQRS library

Adding in the initial concepts of CQRS adaption for MQContract.  Need to implement testing yet

* restructured base

restructured base in preparations for adding CQRS testing

* adding query filtering

added in message filtering for query subscriptions to allow ignoring of some query calls by a given subscription when needed

* extended message filtering

extended message filtering to be defined at a consumer level through interfaces, as well as allow it to exist in cqrs

* moved some shared props

moved some more items to the shared props to improve things for when I move to a new framework and to simplify things

* added testing for consumer extensions

added in testing for the contract consumer connection extensions

* added query filter testing

added in testing for message filtering when attached to a query subscription

* added filtered consumer tests

added in testing for filtered consumers

* added initial cqrs testing

added in initial cqrs testing which currently involves testing the extension calls

* added cqrs unit testing

added in some additional cqrs unit testing as well as expanded building to dotnet10

* completed cqrs testing

completed additional testing for cqrs

* completed commenting

completed all commenting, documentation and testing for 3.0

* Update unittests8x.yml

fixing unit tests action

* sonarqube fixes

correcting some issues reported by sonarqube

* Update AConnection.Consumers.cs

fixing merge conflict issue

* correcting unit tests

correcting some unit test issues

* updated testing for telemetry

added in testing to validate telemetry data and tags

* Update performance.jpg

updating performance stats